Revert wallet loading states
