+++ src/components/Referrals/traders/joinCode/JoinReferralWizard.tsx
@@ -54,7 +54,7 @@ function DiscountPromos() {
 }
 
 export function JoinReferralWizard({ onGoToTraderDashboard }: { onGoToTraderDashboard: () => void }) {
-  const { isConnectModalLoading, openConnectModal } = useConnectModal();
+  const { openConnectModal } = useConnectModal();
   const { isConnected, address } = useAccount();
   const { chainId } = useChainId();
   const history = useHistory();
@@ -97,14 +97,8 @@ export function JoinReferralWizard({ onGoToTraderDashboard }: { onGoToTraderDash
                 <DiscountPromos />
               </div>
               <div className="flex flex-col items-center gap-16">
-                <Button
-                  variant="primary-action"
-                  className="w-full"
-                  type="submit"
-                  disabled={isConnectModalLoading}
-                  onClick={openConnectModal}
-                >
-                  {isConnectModalLoading ? <Trans>Loading...</Trans> : <Trans>Connect wallet</Trans>}
+                <Button variant="primary-action" className="w-full" type="submit" onClick={openConnectModal}>
+                  <Trans>Connect wallet</Trans>
                 </Button>
                 <ExternalLink href={REFERRALS_DOCS_URL} variant="icon" className="text-blue-300">
                   <Trans>Read more</Trans>
