+++ src/components/TokenSelector/ConnectWalletModalContent.tsx
@@ -1,8 +1,6 @@
 import { t, Trans } from "@lingui/macro";
 import { memo } from "react";
 
-import { useConnectModal } from "lib/wallets/useConnectModal";
-
 import ConnectWalletButton from "components/ConnectWalletButton/ConnectWalletButton";
 
 export const ConnectWalletModalContent = memo(function ConnectWalletModalContent({
@@ -12,8 +10,6 @@ export const ConnectWalletModalContent = memo(function ConnectWalletModalContent
   openConnectModal: (() => void) | undefined;
   walletIconUrls: string[] | undefined;
 }) {
-  const { isConnectModalLoading } = useConnectModal();
-
   return (
     <div className="flex grow flex-col items-center justify-center gap-20 p-adaptive">
       {walletIconUrls?.length && (
@@ -40,8 +36,8 @@ export const ConnectWalletModalContent = memo(function ConnectWalletModalContent
           payment options
         </Trans>
       </div>
-      <ConnectWalletButton disabled={isConnectModalLoading} onClick={openConnectModal}>
-        {isConnectModalLoading ? <Trans>Loading...</Trans> : <Trans>Connect wallet</Trans>}
+      <ConnectWalletButton onClick={openConnectModal}>
+        <Trans>Connect wallet</Trans>
       </ConnectWalletButton>
     </div>
   );
