+++ src/lib/wallets/ConnectModalProvider.tsx
@@ -1,31 +1,54 @@
+import { useConnectOrCreateWallet } from "@privy-io/react-auth";
 import { useCallback, useMemo, useState, type ReactNode } from "react";
 
-import { ConnectModalController } from "./ConnectModalController";
+import type { SettlementChainId } from "config/chains";
+import {
+  SELECTED_NETWORK_LOCAL_STORAGE_KEY,
+  SELECTED_NETWORK_WAS_APP_SELECTED_LOCAL_STORAGE_KEY,
+} from "config/localStorage";
+import { isSourceChain } from "config/multichain";
+import { useGmxAccountSettlementChainId } from "context/GmxAccountContext/hooks";
+import { metrics } from "lib/metrics";
+import { switchNetwork } from "lib/wallets";
+
 import { ConnectModalContext } from "./useConnectModal";
 
+function shouldKeepAppSelectedSourceChain(settlementChainId: SettlementChainId) {
+  const rawChainIdFromLocalStorage = localStorage.getItem(SELECTED_NETWORK_LOCAL_STORAGE_KEY);
+  const chainIdFromLocalStorage = rawChainIdFromLocalStorage ? parseInt(rawChainIdFromLocalStorage) : undefined;
+  const selectedNetworkWasAppSelected =
+    localStorage.getItem(SELECTED_NETWORK_WAS_APP_SELECTED_LOCAL_STORAGE_KEY) === "true";
+
+  return selectedNetworkWasAppSelected && isSourceChain(chainIdFromLocalStorage, settlementChainId);
+}
+
 export function ConnectModalProvider({ children }: { children: ReactNode }) {
+  const [settlementChainId] = useGmxAccountSettlementChainId();
   const [connectModalOpen, setConnectModalOpen] = useState(false);
-  const [connectRequestId, setConnectRequestId] = useState(0);
-  const [isConnectModalLoading, setIsConnectModalLoading] = useState(false);
+
+  const handleSuccess = useCallback(() => {
+    setConnectModalOpen(false);
+
+    if (shouldKeepAppSelectedSourceChain(settlementChainId)) {
+      return;
+    }
+
+    void switchNetwork(settlementChainId, true).catch((error) => {
+      metrics.pushError(error, "connectModal.switchNetwork");
+    });
+  }, [settlementChainId]);
+
+  const { connectOrCreateWallet } = useConnectOrCreateWallet({
+    onSuccess: handleSuccess,
+    onError: () => setConnectModalOpen(false),
+  });
 
   const openConnectModal = useCallback(() => {
-    setIsConnectModalLoading(true);
-    setConnectRequestId((currentRequestId) => currentRequestId + 1);
-  }, []);
+    setConnectModalOpen(true);
+    connectOrCreateWallet();
+  }, [connectOrCreateWallet]);
 
-  const value = useMemo(
-    () => ({ openConnectModal, connectModalOpen, isConnectModalLoading }),
-    [openConnectModal, connectModalOpen, isConnectModalLoading]
-  );
+  const value = useMemo(() => ({ openConnectModal, connectModalOpen }), [openConnectModal, connectModalOpen]);
 
-  return (
-    <ConnectModalContext.Provider value={value}>
-      {children}
-      <ConnectModalController
-        requestId={connectRequestId}
-        setConnectModalOpen={setConnectModalOpen}
-        setIsConnectModalLoading={setIsConnectModalLoading}
-      />
-    </ConnectModalContext.Provider>
-  );
+  return <ConnectModalContext.Provider value={value}>{children}</ConnectModalContext.Provider>;
 }
